z3c: Set Wifi country code and regrev to 0
By not setting a concrete value here we leave it up to e.g. wpa_supplicant to interact with the firmware (via nl80211/cfg80211) and pass the device the applicable regulatory hint. By setting these values to zero the device should startup in the generic world regulatory domain, this should prevent a violation of regulatory constraints.

According to Cypress documentation AN214783 / 002-14783 Rev. *E for CYW4339 (=BCM4339):
" The country code (ccode) parameter is for regulatory use and specifies which regulatory tables are to be loaded.
The regulatory revision (regrev) code parameter is for regulatory use and specifies which regulatory tables are to be loaded.
Note: Together, the regrev and ccode parameters set the power limitations necessary to meet the countryspecific regulatory specifications."
Change-Id: Id7fe457b32077997b4459abc5a9be6494f88f622